requirements.txt
# Core dependencies
slskd-api>=0.1.5
python-dotenv>=1.0.0
spotipy>=2.22.0

# Optional dependencies for enhanced functionality
music-tag>=0.4.3

# Telegram Bot integration
python-telegram-bot>=20.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# System dependencies (handled by Dockerfile)
# gcc and g++ are installed via apt-get in Dockerfile
